+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Get correct stock location for intercompany (interco) orders.
 - StockMove : make addresses updatable until it's not realized.
 - Fix a bug where duplicated stock move line were shown in a dashlet.
+- StockMove : Modify Locale for Picking Order report.
 
 ## [5.0.6] - 2018-10-06
 ## Features